add test collector logger
