+++ protostar/composition_root.py
@@ -20,9 +20,10 @@ from protostar.commands.init.project_creator import (
     AdaptedProjectCreator,
     NewProjectCreator,
 )
-from protostar.commands.test.test_result_cli_formatter_visitor import (
-    TestResultCLIFormatter,
+from protostar.commands.test.test_collector_result_logger import (
+    TestCollectorResultLogger,
 )
+from protostar.commands.test.test_result_formatter import TestResultFormatter
 from protostar.compiler import ProjectCairoPathBuilder, ProjectCompiler
 from protostar.migrator import Migrator, MigratorExecutionEnvironment
 from protostar.protostar_cli import ProtostarCLI
@@ -149,8 +150,14 @@ def build_di_container(script_root: Path):
             project_root_path,
             protostar_directory,
             project_cairo_path_builder,
-            test_result_cli_formatter_visitor_builder=TestResultCLIFormatter.Builder(
-                log_color_provider=log_color_provider
+            test_result_cli_formatter=TestResultFormatter(
+                log_color_provider=log_color_provider,
+            ),
+            test_collector_result_logger=TestCollectorResultLogger(
+                logger=logger,
+                test_result_cli_formatter=TestResultFormatter(
+                    log_color_provider=log_color_provider,
+                ),
             ),
         ),
         DeployCommand(gateway_facade, logger),
